# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type ChatInputMode = typeof ChatInputMode[keyof typeof ChatInputMode] | null;


export const ChatInputMode = {
  public: 'public',
  admin: 'admin',
} as const;

export interface ChatInput {
  message: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  hiddenPrompt?: boolean | null;
  /** @nullable */
  mode?: ChatInputMode;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
}

export interface SubElement {
  id: number;
  elementId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order: number;
}

export interface Element {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order: number;
  subElements?: SubElement[];
}

export interface ElementInput {
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order?: number;
}

export interface ElementUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  promptText?: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order?: number;
}

export interface SubElementInput {
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order?: number;
}

export interface SubElementUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  promptText?: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order?: number;
}

export interface ReorderInput {
  ids: number[];
}

export interface SiteSettings {
  heroEyebrow: string;
  heroTitle: string;
  heroSubtitle: string;
  chatPlaceholder: string;
  footerTagline: string;
  footerCopyright: string;
}

export interface SiteSettingsUpdate {
  heroEyebrow?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  chatPlaceholder?: string;
  footerTagline?: string;
  footerCopyright?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface GenerateImageInput {
  /** @minLength 1 */
  prompt: string;
}

export interface GeneratedImage {
  objectPath: string;
}