# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Launch Lemonade API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Chat operations
  - name: elements
    description: Element tile operations
  - name: subelements
    description: Sub-element operations
  - name: siteSettings
    description: Site-wide editable content
  - name: storage
    description: File upload and serving
  - name: images
    description: AI image generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chat:
    post:
      operationId: sendChat
      tags: [chat]
      summary: Send a message to the Launch Lemonade chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /elements:
    get:
      operationId: listElements
      tags: [elements]
      summary: List all element tiles
      responses:
        "200":
          description: List of elements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Element"
    post:
      operationId: createElement
      tags: [elements]
      summary: Create a new element tile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"

  /elements/reorder:
    post:
      operationId: reorderElements
      tags: [elements]
      summary: Reorder element tiles in bulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "204":
          description: Reordered

  /elements/{id}:
    get:
      operationId: getElement
      tags: [elements]
      summary: Get a single element
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Element
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
    patch:
      operationId: updateElement
      tags: [elements]
      summary: Update an element tile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Element"
    delete:
      operationId: deleteElement
      tags: [elements]
      summary: Delete an element tile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /elements/{elementId}/subelements:
    get:
      operationId: listSubElements
      tags: [subelements]
      summary: List sub-elements for an element
      parameters:
        - name: elementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sub-elements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubElement"
    post:
      operationId: createSubElement
      tags: [subelements]
      summary: Create a sub-element
      parameters:
        - name: elementId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubElementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubElement"

  /elements/{elementId}/subelements/{id}:
    patch:
      operationId: updateSubElement
      tags: [subelements]
      summary: Update a sub-element
      parameters:
        - name: elementId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubElementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubElement"
    delete:
      operationId: deleteSubElement
      tags: [subelements]
      summary: Delete a sub-element
      parameters:
        - name: elementId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /site-settings:
    get:
      operationId: getSiteSettings
      tags: [siteSettings]
      summary: Get site-wide editable content
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSiteSettings
      tags: [siteSettings]
      summary: Update site-wide editable content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  /images/generate:
    post:
      operationId: generateTileImage
      tags: [images]
      summary: Generate an image via Gemini and store it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImageInput"
      responses:
        "200":
          description: Generated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        sessionId:
          type: ["string", "null"]
        hiddenPrompt:
          type: ["boolean", "null"]
        mode:
          type: ["string", "null"]
          enum: ["public", "admin", null]

    ChatResponse:
      type: object
      required: [reply, sessionId]
      properties:
        reply:
          type: string
        sessionId:
          type: string

    Element:
      type: object
      required: [id, name, promptText, order]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer
        subElements:
          type: array
          items:
            $ref: "#/components/schemas/SubElement"

    ElementInput:
      type: object
      required: [name, promptText]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer

    ElementUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer

    SubElement:
      type: object
      required: [id, elementId, name, promptText, order]
      properties:
        id:
          type: integer
        elementId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer

    SubElementInput:
      type: object
      required: [name, promptText]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer

    SubElementUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        promptText:
          type: string
        photoUrl:
          type: ["string", "null"]
        linkUrl:
          type: ["string", "null"]
        order:
          type: integer

    ReorderInput:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer

    SiteSettings:
      type: object
      required:
        - heroEyebrow
        - heroTitle
        - heroSubtitle
        - chatPlaceholder
        - footerTagline
        - footerCopyright
      properties:
        heroEyebrow:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        chatPlaceholder:
          type: string
        footerTagline:
          type: string
        footerCopyright:
          type: string

    SiteSettingsUpdate:
      type: object
      properties:
        heroEyebrow:
          type: string
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        chatPlaceholder:
          type: string
        footerTagline:
          type: string
        footerCopyright:
          type: string

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    GenerateImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1

    GeneratedImage:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string