# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatInputMode } from './chatInputMode';

export interface ChatInput {
  message: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  hiddenPrompt?: boolean | null;
  /** @nullable */
  mode?: ChatInputMode;
}