# lib/api-zod/src/generated/types/element.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */
import type { SubElement } from './subElement';

export interface Element {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order: number;
  subElements?: SubElement[];
}