# lib/api-zod/src/generated/types/elementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */

export interface ElementInput {
  name: string;
  /** @nullable */
  description?: string | null;
  promptText: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  order?: number;
}