# lib/api-zod/src/generated/types/generateImageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateImageInput {
  /** @minLength 1 */
  prompt: string;
}