# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Launch Lemonade API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  heroEyebrow: string;
  heroTitle: string;
  heroSubtitle: string;
  chatPlaceholder: string;
  footerTagline: string;
  footerCopyright: string;
}